Vectorfieldvisualization: update to new module structure
