+++ modules/vectorfieldvisualization/include/modules/vectorfieldvisualization/algorithms/integrallineoperations.h
