+++ modules/vectorfieldvisualization/include/modules/vectorfieldvisualization/datastructures/integrallineset.h
