+++ modules/vectorfieldvisualization/src/algorithms/integrallineoperations.cpp
