+++ modules/vectorfieldvisualization/src/datastructures/integrallineset.cpp
