+++ modules/vectorfieldvisualization/src/processors/datageneration/seedpointsfrommask.cpp
