+++ modules/vectorfieldvisualization/src/processors/discardshortlines.cpp
