+++ modules/vectorfieldvisualization/src/processors/integrallinevectortomesh.cpp
