+++ modules/vectorfieldvisualization/src/processors/seedsfrommasksequence.cpp
