+++ modules/vectorfieldvisualization/src/vectorfieldvisualizationmodule.cpp
